feat: implement transformer building blocks including attention layers, encoders, decoders, and positional encoding

+++ src/helpers/helper_math.rs
@@ -1,11 +1,13 @@
-pub fn randn(rng: &mut u64) -> f64 {
+pub fn randn(rng: &mut u64) -> f32 {
     let u1 = 1.0 - lcg_next(rng);
     let u2 = lcg_next(rng);
 
-    (-2.0 * u1.ln()).sqrt() * (2.0 * std::f64::consts::PI * u2).cos()
+    (-2.0 * u1.ln()).sqrt() * (2.0 * std::f32::consts::PI * u2).cos()
 }
 
-fn lcg_next(state: &mut u64) -> f64 {
-    *state = state.wrapping_mul(6364136223846793005).wrapping_add(1442695040888963407);
-    (*state >> 11) as f64 / (1u64 << 53) as f64
+fn lcg_next(state: &mut u64) -> f32 {
+    *state = state
+        .wrapping_mul(6364136223846793005)
+        .wrapping_add(1442695040888963407);
+    (*state >> 11) as f32 / (1u64 << 53) as f32
 }
