+++ src/tensor/array.rs
@@ -1,13 +1,13 @@
 pub struct Array<const ArrSize: usize> {
-    pub data: [f64; ArrSize],
+    pub data: [f32; ArrSize],
 }
 
 impl<const ArrSize: usize> Array<ArrSize> {
-    pub fn new(data: [f64; ArrSize]) -> Self {
+    pub fn new(data: [f32; ArrSize]) -> Self {
         Self { data }
     }
 
-    pub fn from_arr(data: [f64; ArrSize]) -> Self {
+    pub fn from_arr(data: [f32; ArrSize]) -> Self {
         Self { data }
     }
 }
